Move SystemPropertyReferencesTestCase to core-model-test
